% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_bands.R
\name{as.calibrationband}
\alias{as.calibrationband}
\title{use method}
\usage{
as.calibrationband(x, ...)
}
\arguments{
\item{x}{object of class isoband}

\item{...}{further arguments to be passed to or from methods;}
}
\description{
use method
}
