% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_bands.R
\name{calibration_bands}
\alias{calibration_bands}
\title{Confidence bands for monotone probabilities}
\usage{
calibration_bands(
  x,
  y,
  alpha = 0.05,
  method = "standard",
  digits = NULL,
  nc = FALSE
)
}
\arguments{
\item{x}{covariate.}

\item{y}{response variable (in {0,1}).}

\item{alpha}{type one error probability (1 minus the confidence level).}

\item{method}{\code{"standard"} for the original method proposed in the
article, \code{"nc"} for the non-crossing variant, \code{"round"} for
rounding the covariate, or \code{"YB"} for the bounds by Yang & Barber
(2019).}

\item{digits}{number of digits for method \code{"round"}. Default is 2.
Has no effect for the other methods.}

\item{nc}{use non-crossing bands for method \code{"standard"} or
\code{"round"}. Has no effect for method \code{"YB"}. Default is
\code{FALSE}.}
}
\value{
An object of class \code{isoband}, which is a list containing

\item{\code{x_original,y_original}}{the \code{x} and \code{y} supplied for
    computing the confidence band.}

\item{\code{alpha}}{the given type one error probability (1 minus the nominal
    coverage of the band).}

\item{\code{method}}{the selected method for computing the band.}

\item{\code{nc}}{the selected method for non-crossing.}

\item{\code{digits}}{the given digits for method \code{"round"}
    (or \code{NULL} for method \code{"standard"}).}

\item{\code{x}}{the position on the x-axis of the lower and upper bound (see
    last item).}

\item{\code{isoy}}{isotonic regression of \code{y} at points \code{x}.}

\item{\code{lwr,upr}}{the lower and upper bound, for each value of \code{x}.
    The upper bound extends to the left and the lower bound to the right,
    that is, the upper bound for \code{x[i]<s<x[i+1]} is \code{upr[i+1]},
    and the lower bound for \code{x[i]<s<x[i+1]} is \code{lwr[i]}.}
}
\description{
Confidence bands for monotone probabilities
}
\author{
Alexander Henzi
}
