% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate}
\alias{interpolate}
\title{Interpolation (required in isobands)}
\usage{
interpolate(x, y, xout, right = 1)
}
\arguments{
\item{x}{position on x axis.}

\item{y}{position on y axis.}

\item{xout}{positions to interpolate.}

\item{right}{if \code{1}, the interpolation at xout with
x[i] < xout <= x[i+1] is y[i+1], if \code{0} it is y[i].}
}
\description{
Interpolation (required in isobands)
}
