% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p.dat}
\alias{p.dat}
\title{Plot data for bands (assures correct connecting point)}
\usage{
p.dat(bands)
}
\arguments{
\item{bands}{object bands from calibration_band function}
}
\description{
Plot data for bands (assures correct connecting point)
}
