% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.calibrationband}
\alias{plot.calibrationband}
\title{Plotting monotone confidence bands}
\usage{
\method{plot}{calibrationband}(x, ..., plot = FALSE, add = FALSE, col = "black")
}
\arguments{
\item{x}{output of the \code{calibrationband} functions.}

\item{...}{further arguments to be passed to or from methods;}

\item{plot}{if \code{TRUE}, a simple plot of the confidence bands is
displayed (default \code{FALSE}).}

\item{add}{if \code{TRUE}, adds the bands to an existing plot (default
\code{FALSE}).}

\item{col}{color of bands.}
}
\value{
A list containing the \code{x,y} to be called in the function \code{plot} to
plot the confidence band, separated for the upper and lower bound (returned
invisibly).
}
\description{
Plotting monotone confidence bands
}
\details{
When plotting the monotone confidence band, the upper bound should be
extended to the left, that is, the bound at \code{x[i]} is valid on the
interval \code{(x[i-1],x[i]]}. The lower bound should be extended to the
right, i.e. the bound at x[i] is extended to the interval \code{[x[i],x[i +
1])}. This function creates x and y values for correct plotting of these
bounds.
}
\author{
Alexander Henzi
}
