% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.calibrationband}
\alias{print.calibrationband}
\title{Print monotone confidence bands}
\usage{
\method{print}{calibrationband}(x, ...)
}
\arguments{
\item{x}{object of class calibrationband}

\item{...}{further arguments to be passed to or from methods;}
}
\description{
Print monotone confidence bands
}
