% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/summary.R
\name{print.summary.calibrationband}
\alias{print.summary.calibrationband}
\alias{summary.calibrationband}
\title{print summary}
\usage{
\method{print}{summary.calibrationband}(x, ...)

\method{summary}{calibrationband}(object, ...)
}
\arguments{
\item{x}{summary object}

\item{...}{further arguments to be passed to or from methods;}

\item{object}{object of class calibrationband}
}
\description{
print summary

summarize isoband output
}
