% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.calibrationband}
\alias{summary.calibrationband}
\title{summarize calibration band  object}
\usage{
\method{summary}{calibrationband}(object, ...)
}
\arguments{
\item{object}{object of class \code{calibrationband}}

\item{...}{further arguments to be passed to or from methods;}
}
\value{
A \code{'summary.reliability'} object, which is also a
tibble (see \code{\link[tibble:tibble]{tibble::tibble()}}) with columns:
\tabular{ll}{
min_x \tab minimal x-coordinate of  misscalibration segemnt (ordered by length). \cr
max_x \tab maximal x-coordinate of  misscalibration segemnt (ordered by length).
}
}
\description{
An object of class \code{calibrationband} contains the calibration band
coordinates, the pairs of original observation and  forecast values,
and the recalibrated forecasts given by isotonic regression.
The function \code{summary.reliabilitydiag} calculates the areas of
miscalibration.
}
