% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.calibrationband}
\alias{autoplot.calibrationband}
\title{Plotting monotone confidence bands}
\usage{
\method{autoplot}{calibrationband}(
  object,
  ...,
  approx.equi = F,
  points = 500,
  isoreg = T,
  shaddow = "blue",
  diag = "red",
  cut.bands = FALSE
)
}
\arguments{
\item{object}{object of class calibrationband}

\item{...}{further arguments to be passed to or from methods;}

\item{approx.equi}{If \code{NULL}, the bands are drawn for each
prediction-realization pair. In large data sets this might result in capacity
consuming plots. Other choices than \code{NULL} result in equidistant
interpolation at by default 700 \code{points}. See bellow.}

\item{points}{Interpolate at \code{points} equidistant values between 0 and 1
(if \code{cut.bands=F}) or between the most extreme prediction values
(if \code{cut.bands=T}). Note, for smooth accurate illustration in transition areas
(changes between miscalibrated and calibrated areas) we add those critical
additional points the initial value of  \code{points}.}

\item{isoreg}{If \code{NULL}, raws the isotonic/monotonic non-parametric
regression curve.}

\item{shaddow}{If \code{NULL}, raws a blueish ribbon.}

\item{diag}{If default, a plain, black diagonal line between the coordinates
0,0 and 1,1 is draw. Otherwise, the diagonal line is colored
(red if miscalibrated, gray if calibrated).
Note, if the latter is selected and if \code{cut.bands=T}, only the area in
between the minimal and maximal prediction value is colored.}

\item{cut.bands}{Cut the bands at most extreme prediction values. Bands will not be extended to 0 and 1 respectively.}
}
\description{
Plotting monotone confidence bands
}
