% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.calibrationband}
\alias{print.calibrationband}
\alias{print.summary.calibrationband}
\title{Print monotone confidence bands}
\usage{
\method{print}{calibrationband}(x, ...)

\method{print}{summary.calibrationband}(x, ..., n = 3)
}
\arguments{
\item{x}{object of class \code{calibrationband}}

\item{...}{Further arguments to be passed to or from methods;
in particular these passed to \code{\link{autoplot.calibrationband}}}

\item{n}{number of areas of miscalibration printed.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Printing methods for \code{'calibrationband'} and
\code{'summary.calibrationband'} objects.
}
\details{
\code{print.calibrationband} always sends an autoplot object to the
current graphics device and prints a summary to the console.
}
\seealso{
\code{\link{autoplot.calibrationband}},
  \code{\link{summary.calibrationband}}
}
