% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.calibrationband}
\alias{summary.calibrationband}
\title{summarize calibration band  object}
\usage{
\method{summary}{calibrationband}(object, ...)
}
\arguments{
\item{object}{object of class \code{calibrationband}}

\item{...}{Further arguments to be passed to or from methods.}
}
\value{
A \code{'summary.reliability'} object, which is also a
tibble (see \code{\link[tibble:tibble]{tibble::tibble()}}) with columns:
\tabular{ll}{
min_x \tab minimal x-coordinate of  misscalibration segment (ordered by length). \cr
max_x \tab maximal x-coordinate of  misscalibration segment (ordered by length).
}
}
\description{
An object of class \code{calibrationband} contains the calibration band
coordinates, the pairs of original observation and forecast values,
and the recalibrated forecasts obtained by isotonic regression.
The function \code{summary.reliabilitydiag} calculates the areas of
miscalibration.
}
\examples{
set.seed(123)
s=.8
n=10000
x <- sort(runif(n))

p <- function(x,s){p = 1/(1+((1/x*(1-x))^(s+1)));return(p)}
dat <- data.frame(pr=x, y=rbinom(n,1,p(x,s)))

cb <- calibration_bands(x=dat$pr, y=dat$y,alpha=0.05, method="round", digits =3)

summary(cb)
print(summary(cb), n=5)

}
